requirements.txt
APScheduler==3.7.0
azure-common==1.1.27
azure-core==1.13.0
azure-identity==1.5.0
azure-mgmt-compute==20.0.0
azure-mgmt-core==1.2.2
azure-mgmt-network==18.0.0
boto3==1.17.51
botocore==1.20.51
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cryptography==3.4.7
DateTime==4.3
docker==5.0.0
Flask==2.0.0
idna==2.10
isodate==0.6.0
itsdangerous==2.0.0
Jinja2==3.0.0
jmespath==0.10.0
MarkupSafe==2.0.0
msal==1.11.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.0
portalocker==1.7.1
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.7
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
websocket-client==0.59.0
Werkzeug==2.0.0
zope.interface==5.4.0
